+++ src/helpers/ERC721Proof.ts
@@ -1,4 +1,3 @@
-import { Entropy } from 'entropy-string'
 import { utils } from 'ethers'
 import BalanceSignature from 'models/BalanceSignature'
 import BaseProof from 'helpers/BaseProof'
@@ -7,6 +6,7 @@ import Proof from 'models/Proof'
 import ProofResult from 'models/ProofResult'
 import PublicKey from 'models/PublicKey'
 import Signature from 'models/Signature'
+import generateR2AndS2 from 'helpers/generateR2AndS2'
 import unpackSignature from 'helpers/unpackSignature'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -109,17 +109,7 @@ export default class ERC721Proof
     balanceSignature: BalanceSignature,
     eddsaPublicKey: PublicKey
   ) {
-    const hexadecimalCharset = '0123456789ABCDEF'
-    const randomHexadecimalNumber1 = new Entropy({
-      charset: hexadecimalCharset,
-      bits: 64,
-    }).string()
-    const randomHexadecimalNumber2 = new Entropy({
-      charset: hexadecimalCharset,
-      bits: 64,
-    }).string()
-    const r2 = BigInt('0x' + randomHexadecimalNumber1).toString(10)
-    const s2 = BigInt('0x' + randomHexadecimalNumber2).toString(10)
+    const { r2, s2 } = generateR2AndS2()
     const addressInputs = await this.inputsForSignature(
       eddsaPublicKey,
       ownershipSignature,
